item model

+++ lib/todoable/list.rb
@@ -1,28 +1,32 @@
 module Todoable
   class List
-    ENDPOINT = 'lists'
+    ENDPOINT = 'lists'.freeze
 
     class << self
       def all
         Todoable.http.get(ENDPOINT).body['lists'].map do |params|
-          new(params['name'], params['id'], params['src'])
+          from_params params
         end
       end
 
-      def create(name)
-        raise Todoable::Error if name.nil? || name.empty?
+      def from_params(params)
+        list = new params['name'], params['id'], params['src']
+      end
 
-        Todoable.http.post(ENDPOINT, list: { name: name })
+      def create(name)
+        result = Todoable.http.post(ENDPOINT, list: { name: name })
+        raise UnprocessableError, result.body if result.status == 422
+
+        from_params result.body
       end
 
       def find(id)
-        Todoable.http.get([ENDPOINT,id].join('/')).body['list']
+        from_params Todoable.http.get([ENDPOINT, id].join('/')).body['list']
       end
     end
 
     attr_reader :name
-    attr_accessor :items
-    def initialize(name, id, src)
+    def initialize(name, id = nil, src = nil)
       @items = []
       @name = name
       @id = id
@@ -36,11 +40,28 @@ module Todoable
       update
     end
 
+    def items
+      result = Todoable.http.get(@src)
+      @name = result.body['name']
+      Item.from_array result.body['items']
+    end
+
+    def add_item(name)
+      Todoable.http.post([@src, 'items'].join('/'), item: { name: name }).body['item']
+    end
+
     def destroy
-      Todoable.http.delete(@src)
+      return if @src.nil?
+
+      result = Todoable.http.delete(@src)
+      raise NotFound if result.status == 404
+
+      @id = nil
+      @src = nil
     end
 
     private
+
     private_class_method :find
 
     def refresh
